requirements.txt
spotipy
pyyaml
pillow
numpy